+++ gDRplots/R/plot-pheatmap.R
@@ -15,7 +15,9 @@
 #'    the last one - to the max; for "x_std" - that will be reversed
 #' @param no_breaks numeric number of breaks on scale
 #' @param cluster_rows logical flag whether rows should be clustered;
-#'   the dendrogram will not be shown for the matrix with any NA value
+#'   the dendrogram will not be shown for the matrix with any dimension greater than 200.
+#' @param distfun function used to compute the distance (dissimilarity) between rows;
+#'   used for the dendrogram when \code{cluster_rows} is set to TRUE
 #' @param lbl_by_CellLineName logical flag whether heatmap should be described by CellLineNames instead of clid
 #' @param lbl_by_DrugName logical flag whether heatmap should be described by DrugName instead of Gnumber
 #' 
@@ -63,6 +65,7 @@ pheatmap_qc <- function(
     colors_vec = c("black", "grey99"),
     no_breaks = 50,
     cluster_rows = TRUE,
+    distfun = compute_distances,
     lbl_by_CellLineName = FALSE,
     lbl_by_DrugName = FALSE) {
   
@@ -75,6 +78,7 @@ pheatmap_qc <- function(
   stopifnot("Must be a valid color name" = all(vapply(colors_vec, is_valid_color, logical(1))))
   checkmate::assert_int(no_breaks, lower = 2)
   checkmate::assert_flag(cluster_rows)
+  checkmate::assert_function(distfun)
   checkmate::assert_flag(lbl_by_CellLineName)
   checkmate::assert_flag(lbl_by_DrugName)
   zero_conc_scaling_factor <- 
@@ -172,12 +176,11 @@ pheatmap_qc <- function(
   names(drug_annotation_colors) <- drug_to_colored
   
   # dendrogram
-  cluster_rows <- if (cluster_rows && !any(is.na(mat_cvd)) && NROW(mat_cvd) >= 2) {
-    stats::hclust(stats::dist(mat_cvd))
-  } else {
-    FALSE
-  }
-  
+  gDR_cluster_condition <- any(dim(mat_cvd) < 200)  # gDR standard
+  cluster_rows <- .pheatmap_cluster_param(mat_to_cluster = mat_cvd,
+                                          cluster_flag = cluster_rows,
+                                          distfun = distfun,
+                                          additional_condition = gDR_cluster_condition)
   # heatmap labels
   if (lbl_by_CellLineName) {
     row_lbls <- tab_response[, unique(.SD), .SDcols = c(cellline_name, clid)][order(rownames(mat_cvd))]
@@ -255,9 +258,11 @@ pheatmap_qc <- function(
 #'   are matched using corresponding names from the required  \code{DrugName} column.
 #'   Note that color schemes takes into account if variable is continuous or discrete.
 #' @param cluster_rows logical flag whether rows should be clustered;
-#'   the dendrogram will not be shown for the matrix with any NA values
+#'   the dendrogram will not be shown for the matrix with any dimension greater than 200.
 #' @param cluster_cols logical flag whether columns should be clustered;
-#'   the dendrogram will not be shown for the matrix with any NA values or -Inf/Inf value
+#'   the dendrogram will not be shown for the matrix with any dimension greater than 200.
+#' @param distfun function used to compute the distance (dissimilarity) between both rows and columns;
+#'   used for the dendrogram when \code{cluster_rows} or \code{cluster_cols} is set to TRUE
 #' @param annotation_col \code{data.table} that specifies the annotations shown above the heatmap.
 #'   Each row defines the features for a specific column. The columns in the data and in the annotation
 #'   are matched using corresponding names from the required  \code{CellLineName} column.
@@ -343,6 +348,7 @@ pheatmap_with_anno_sa <- function(
     no_breaks = 50,
     cluster_rows = TRUE,
     cluster_cols = TRUE,
+    distfun = compute_distances,
     annotation_row = NULL,
     annotation_col = NULL,
     annotation_colors = NULL) {
@@ -360,6 +366,7 @@ pheatmap_with_anno_sa <- function(
   checkmate::assert_int(no_breaks, lower = 2)
   checkmate::assert_flag(cluster_rows)
   checkmate::assert_flag(cluster_cols)
+  checkmate::assert_function(distfun)
   checkmate::assert_data_table(annotation_row, null.ok = TRUE)
   if (!is.null(annotation_row)) {
     checkmate::assert_names(names(annotation_row), must.include = drug_name)
@@ -461,18 +468,15 @@ pheatmap_with_anno_sa <- function(
   t_mat_cvd[] <- vapply(t_mat_cvd, function(x) qmfun(x), numeric(1))
   
   # dendrogram
-  cluster_condition <- !any(is.na(t_mat_cvd)) && !any(is.infinite(t_mat_cvd)) && 
-    any(dim(t_mat_cvd) < 200) # gDR standard
-  cluster_rows <- if (cluster_rows && cluster_condition && NROW(t_mat_cvd) >= 2) {
-    stats::hclust(stats::dist(t_mat_cvd))
-  } else {
-    FALSE
-  }
-  cluster_cols <- if (cluster_cols && cluster_condition && NCOL(t_mat_cvd) >= 2) {
-    stats::hclust(stats::dist(t(t_mat_cvd)))
-  } else {
-    FALSE
-  }
+  gDR_cluster_condition <- any(dim(t_mat_cvd) < 200)  # gDR standard
+  cluster_rows <- .pheatmap_cluster_param(mat_to_cluster = t_mat_cvd,
+                                          cluster_flag = cluster_rows,
+                                          distfun = distfun,
+                                          additional_condition = gDR_cluster_condition)
+  cluster_cols <- .pheatmap_cluster_param(mat_to_cluster = t(t_mat_cvd),
+                                          cluster_flag = cluster_cols,
+                                          distfun = distfun,
+                                          additional_condition = gDR_cluster_condition)
   
   # prep hm color palette
   min_val <- min(t_mat_cvd[!is.infinite(t_mat_cvd)], na.rm = TRUE)
@@ -536,9 +540,9 @@ pheatmap_with_anno_sa <- function(
 #'   \code{DrugName_2} and \code{Concentration_2} columns.
 #'   Note that color schemes takes into account if the variable is continuous or discrete.
 #' @param cluster_rows logical flag whether indicating rows should be clustered;
-#'   the dendrogram will not be shown for the matrix with any NA values
+#'   the dendrogram will not be shown for the matrix with any dimension greater than 200.
 #' @param cluster_cols logical flag indicating whether columns should be clustered;
-#'   the dendrogram will not be shown for the matrix with any NA values or -Inf/Inf value
+#'   the dendrogram will not be shown for the matrix with any dimension greater than 200.
 #' @param annotation_col \code{data.table} that specifies the annotations shown above the heatmap.
 #'   Each row defines the features for a specific column. The columns in the data and in the annotation
 #'   are matched using corresponding names from the required  \code{CellLineName} column.
@@ -626,6 +630,7 @@ pheatmap_with_anno_cd <- function(
     no_breaks = 50,
     cluster_rows = TRUE,
     cluster_cols = TRUE,
+    distfun = compute_distances,
     annotation_row = NULL,
     annotation_col = NULL,
     annotation_colors = NULL) {
@@ -645,6 +650,7 @@ pheatmap_with_anno_cd <- function(
   checkmate::assert_int(no_breaks, lower = 2)
   checkmate::assert_flag(cluster_rows)
   checkmate::assert_flag(cluster_cols)
+  checkmate::assert_function(distfun)
   checkmate::assert_data_table(annotation_row, null.ok = TRUE)
   if (!is.null(annotation_row)) {
     checkmate::assert_names(names(annotation_row), must.include = c(drug_name, drug_name_2, conc_2))
@@ -753,18 +759,15 @@ pheatmap_with_anno_cd <- function(
   t_mat_cvd[] <- vapply(t_mat_cvd, function(x) purrr::quietly(qmfun)(x)$result, numeric(1))
   
   # dendrogram
-  cluster_condition <- !any(is.na(t_mat_cvd)) && !any(is.infinite(t_mat_cvd)) && 
-    any(dim(t_mat_cvd) < 200) # gDR standard
-  cluster_rows <- if (cluster_rows && cluster_condition && NROW(t_mat_cvd) >= 2) {
-    stats::hclust(stats::dist(t_mat_cvd))
-  } else {
-    FALSE
-  }
-  cluster_cols <- if (cluster_cols && cluster_condition && NCOL(t_mat_cvd) >= 2) {
-    stats::hclust(stats::dist(t(t_mat_cvd)))
-  } else {
-    FALSE
-  }
+  gDR_cluster_condition <- any(dim(t_mat_cvd) < 200)  # gDR standard
+  cluster_rows <- .pheatmap_cluster_param(mat_to_cluster = t_mat_cvd,
+                                          cluster_flag = cluster_rows,
+                                          distfun = distfun,
+                                          additional_condition = gDR_cluster_condition)
+  cluster_cols <- .pheatmap_cluster_param(mat_to_cluster = t(t_mat_cvd),
+                                          cluster_flag = cluster_cols,
+                                          distfun = distfun,
+                                          additional_condition = gDR_cluster_condition)
   
   # prep hm color palette
   min_val <- min(t_mat_cvd, na.rm = TRUE)
@@ -900,6 +903,7 @@ pheatmap_with_anno_combo <- function(
     no_breaks = 50,
     cluster_rows = TRUE,
     cluster_cols = TRUE,
+    distfun = compute_distances,
     annotation_row = NULL,
     annotation_col = NULL,
     annotation_colors = NULL) {
@@ -917,6 +921,7 @@ pheatmap_with_anno_combo <- function(
   checkmate::assert_int(no_breaks, lower = 2)
   checkmate::assert_flag(cluster_rows)
   checkmate::assert_flag(cluster_cols)
+  checkmate::assert_function(distfun)
   checkmate::assert_data_table(annotation_row, null.ok = TRUE)
   if (!is.null(annotation_row)) {
     checkmate::assert_names(names(annotation_row), must.include = c(drug_name, drug_name_2))
@@ -1255,4 +1260,60 @@ fill_ann_color_map <- function(dt_ann,
     }
   }
   return(map_ann)
+}
+
+#' Compute value of param cluster_rows or cluster_cols in pheatmap::pheatmap
+#' 
+#' The \code{cluster_rows} and \code{cluster_cols} parameters  pheatmap::pheatmap may take values:
+#' - boolean determining if rows/columns should be clustered
+#' - \code{hclust} object,
+#' this function allows to calculate proper value taking into account matrix, additional condition and 
+#' selected function used to compute the distance in \code{hclust} object
+#' 
+#' To compute the correct value when clustering columns - use the transposed source matrix as \code{mat_to_cluster}
+#'
+#' @param mat_to_cluster numeric matrix to be clustered; cluster dimension must be named
+#' @param cluster_flag logical flag whether rows/should be clustered;
+#' @param distfun function used to compute the distance (dissimilarity) between rows;
+#'   defaults to \code{\link[stats]{dist}}
+#' @param additional_condition additional logical flag whether rows/columns should be clustered (will be included 
+#'   as the logical sum of \code{cluster_flag} and \code{additional_condition})
+#'
+#' @return logical flag determining if rows should be clustered or \code{hclust} object.
+#' 
+#' @seealso \code{\link{compute_distances}}
+#' 
+#' @examples
+#' mat <- matrix(1:24, nrow = 4)
+#' rownames(mat) <- sprintf("row_%s", 1:4)
+#' colnames(mat) <- sprintf("col_%s", 1:6)
+#' .pheatmap_cluster_param(mat)
+#' .pheatmap_cluster_param(t(mat)) |> plot()
+#' .pheatmap_cluster_param(t(mat), distfun = compute_distances) |> plot()
+#' 
+#' mat[2,2] <- NA
+#' mat[2,1] <- Inf
+#' .pheatmap_cluster_param(mat)
+#' .pheatmap_cluster_param(mat, distfun = compute_distances)
+#' .pheatmap_cluster_param(t(mat), distfun = compute_distances)
+#' add_cond <- NCOL(mat) > 10
+#' .pheatmap_cluster_param(mat, distfun = compute_distances, additional_condition = add_cond)
+#' 
+#' @keywords internal
+.pheatmap_cluster_param <- function(mat_to_cluster,
+                                    cluster_flag = TRUE,
+                                    distfun = stats::dist,
+                                    additional_condition = TRUE) {
+  
+  checkmate::assert_flag(cluster_flag)
+  checkmate::assert_matrix(mat_to_cluster, mode = "numeric", row.names = "unique")
+  checkmate::assert_function(distfun)
+  checkmate::assert_flag(additional_condition)
+  
+  if (cluster_flag && additional_condition && NROW(mat_to_cluster) >= 2) {
+    tryCatch(stats::hclust(distfun(mat_to_cluster)),
+             error = function(e) return(FALSE)) # if any problem with distfun -> no clustering
+  } else {
+    FALSE
+  }
 }